+++ src/display/mod.rs
@@ -1,0 +1,2 @@
+mod styling;
+mod views;
