+++ src/events/handlers.rs
@@ -1,5 +1,5 @@
 use crate::{
-    application::ApplicationEvent,
+    application::{ApplicationEvent, GUI},
     directory::{self, Directory},
 };
 use async_tungstenite::tungstenite;
@@ -18,4 +18,31 @@ use std::{
 pub async fn on_scan_start(file_path: PathBuf, recv: Receiver<()>) -> Directory {
     let dir = directory::read_dir(&file_path, &recv);
     dir
+}
+
+pub fn on_scan_request(app: &mut GUI) -> Command<ApplicationEvent> {
+    {
+        app.scanning = true;
+        app.pressed_cancel = false;
+        match app.selected_drive.clone() {
+            Some(drive) => {
+                let (send, recv) = channel();
+                app.cancel_sender = Some(send);
+
+                let selected_path: PathBuf =
+                    // PathBuf::from(String::from(r"C:\Users\AJ\Desktop"));
+                app.paths.get(&drive).expect("Letter not found").clone();
+
+                let (send, recv) = channel();
+                Command::perform(
+                    handlers::on_scan_start(selected_path, recv),
+                    ApplicationEvent::ScanFinished,
+                )
+            }
+            None => {
+                println!("No drive selected");
+                Command::none()
+            }
+        }
+    }
 }
