+++ src/logic/analyzer.rs
