+++ src/logic/mod.rs
@@ -1,0 +1,2 @@
+mod analyzer;
+mod directory;
